tweetのポリシーを作成

+++ app/Http/Controllers/Api/TweetController.php
@@ -8,6 +8,7 @@ use App\Models\Tweet;
 use App\Services\TweetService;
 use App\Http\Requests\StoreTweetRequest;
 use App\Http\Requests\UpdateTweetRequest;
+use Illuminate\Support\Facades\Gate;
 
 class TweetController extends Controller
 {
@@ -23,6 +24,7 @@ class TweetController extends Controller
    */
   public function index()
   {
+    Gate::authorize('viewAny', Tweet::class);
     $tweets = $this->tweetService->allTweets();
     return response()->json($tweets);
   }
@@ -32,6 +34,7 @@ class TweetController extends Controller
    */
   public function store(StoreTweetRequest $request)
   {
+    Gate::authorize('create', Tweet::class);
     $tweet = $this->tweetService->createTweet($request->only('tweet'), $request->user());
     return response()->json($tweet, 201);
   }
@@ -41,6 +44,7 @@ class TweetController extends Controller
    */
   public function show(Tweet $tweet)
   {
+    Gate::authorize('view', $tweet);
     return response()->json($tweet);
   }
 
@@ -49,6 +53,7 @@ class TweetController extends Controller
    */
   public function update(UpdateTweetRequest $request, Tweet $tweet)
   {
+    Gate::authorize('update', $tweet);
     $updatedTweet = $this->tweetService->updateTweet($tweet, $request->all());
 
     return response()->json($tweet);
@@ -59,6 +64,7 @@ class TweetController extends Controller
    */
   public function destroy(Tweet $tweet)
   {
+    Gate::authorize('delete', $tweet);
     $this->tweetService->deleteTweet($tweet);
     return response()->json(['message' => 'Tweet deleted successfully']);
   }
